requirements.txt
streamlit>=1.35.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
torchvision